+++ source/common/modules/markdown-editor/util/configuration.ts
@@ -78,6 +78,7 @@ export interface EditorConfiguration {
   showStatusbar: boolean
   showFormattingToolbar: boolean
   darkMode: boolean
+  darkModeEditor: 'match'|'light'|'dark'
   theme: MarkdownTheme
   margins: 'S'|'M'|'L'
   highlightWhitespace: boolean
@@ -144,6 +145,7 @@ export function getDefaultConfig (): EditorConfiguration {
     showStatusbar: false,
     showFormattingToolbar: true,
     darkMode: false,
+    darkModeEditor: 'match',
     theme: 'berlin',
     margins: 'M',
     highlightWhitespace: false,