+++ test/localise-number.spec.ts
@@ -13,12 +13,8 @@
  * END HEADER
  */
 
-// DEBUG: This test is currently failing since mocha is unable to load ES6 modules.
-// It all began with bcp-47 being converted to an ES6 module, and it's preferable
-// to use import instead of require either way, so we're gonna disable this test
-// for the time being. Plus, the test never failed, and I don't see any reason
-// why we would need to change that function, so ... just don't change it, and
-// we'll be good!
+// BUG: This test essentially just tests the built-in fallback to the function,
+// since it will not be able to load i18n data from main.
 import localiseNumber from '@common/util/localise-number'
 import assert from 'assert'
 
@@ -32,17 +28,9 @@ const localiseNumberTesters = [
 ]
 
 describe('Utility#localiseNumber()', function () {
-  before(function () {
-    global.i18n = global.i18nFallback = { 'localise': { 'thousand_delimiter': ',', 'decimal_delimiter': '.' } }
-  })
-
   for (let test of localiseNumberTesters) {
     it(`should return ${test.expected} for ${test.input.toString()}`, function () {
       assert.strictEqual(localiseNumber(test.input), test.expected)
     })
   }
-
-  after(function () {
-    global.i18n = global.i18nFallback = undefined // Unset
-  })
 })